backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Ragster"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.1",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.0",
]
